Add files via upload
A new project. Inspired by RuneScape (https://www.runescape.com) and Old School RuneScape Discord Bot from GC (https://github.com/oldschoolgg/oldschoolbot). So far I did get the data for the trees (exp, levels, log types). I did a JSON-file which stores the skill exps. Chopping function that prints the data to us and a xp function that gives the experiences for us. And also a level function that checks if the experience is more than a set point, if it is it will give us one level, and it will start to check again. The main file will be the file where every skill will be moved to, Mining, Agility etc. Let's see where this gets, probably not far but I want to have fun and learn at the sametime!